configuration/requirements.txt
asn1crypto==0.24.0
cffi==1.11.5
cryptography==2.2.2
Django==2.0.5
django-environ==0.4.4
django-rest-knox==3.1.4
django-webpack-loader==0.6.0
djangorestframework==3.8.2
environ==1.0
idna==2.7
psycopg2==2.7.5
psycopg2-binary==2.7.5
pycparser==2.18
pyOpenSSL==18.0.0
pytz==2018.4
six==1.11.0
whitenoise==3.3.1
